Cambié int por float en la función de matrices_yordan. Primera versión de matrices_carlo, generación y llenado de matrices sin numpy

+++ matrices_yordan.py
@@ -8,7 +8,7 @@ def llenado():
     for i in range(m):
         f=[]
         for j in range(n):
-            f.append(int(input(f'Ingrese el elemento {(i+1,j+1)}:')))
+            f.append(float(input(f'Ingrese el elemento {(i+1,j+1)}:')))
         A.append(f)
     
     return A
